requirements.txt
anthropic>=0.49.0
python-dotenv>=1.0.0
typing-extensions>=4.10.0
types-boto3>=1.0.2
boto3-stubs[dynamodb]>=1.34.0